add condn to receive msg from slack.
Add condition when receiving message from slack.

Slack sents an initial HTTP Post request while registering a new path as an event handler. While sending that request, slack won't sent any text message in it and we also don't want to invoke frontend in that request.

Using If block for bypassing that request.